Move code from ImportData to OutdatedNotificationJob and merge two similar tests
